Moving code around to be more logical now that we have SSO based auth proven
